moved all the commands into another file, clean some comments.
