+++ src/server.cpp
@@ -6,7 +6,7 @@
 /*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/18 15:41:53 by gyong-si          #+#    #+#             */
-/*   Updated: 2025/05/18 11:25:32 by gyong-si         ###   ########.fr       */
+/*   Updated: 2025/05/19 10:54:53 by gyong-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -217,10 +217,6 @@ void	Server::handleIncomingNewClient()
 	// this adds the client into the clients list
 	_clients.push_back(newClient);
 
-	// set the client fd to non blocking
-	//int flags = fcntl(client_fd, F_GETFL, 0);
-	//fcntl(client_fd, F_SETFL, flags | O_NONBLOCK);
-
 	// this adds the client fd into epoll
 	struct epoll_event clientEvent;
 	clientEvent.events = EPOLLIN;
@@ -233,72 +229,6 @@ void	Server::handleIncomingNewClient()
 	}
 }
 
-/**
- * handles the client's password and checks if it matches with the server.
- */
-void Server::handlePass(int fd, std::list<std::string> cmd_list)
-{
-	Client *client = getClientByFd(fd);
-	if (!client)
-		return ;
-	if (cmd_list.size() != 2)
-	{
-		// need to change this?
-		send(fd, "ERROR :PASS command requires exactly one argument\r\n", 50, 0);
-		return ;
-	}
-	// iterate to the second item in the list
-	std::list<std::string>::const_iterator it = cmd_list.begin();
-	++it;
-	const std::string &provided = *it;
-	if (_password == sha256(provided))
-	{
-		client->authenticate();
-		send(fd, "NOTICE AUTH :Password accepted\r\n", 33, 0);
-		std::cout << "Client " << fd << " : has been authenticated.\n";
-	}
-	else
-		send(fd, "ERROR :Invalid password\r\n", 26, 0);
-}
-
-void Server::handleNick(int fd, std::list<std::string> cmd_list)
-{
-	Client *client = getClientByFd(fd);
-	if (!client)
-		return ;
-	if (cmd_list.size() != 2)
-	{
-		// I need to change this
-		send(fd, "ERROR :PASS command requires exactly one argument\r\n", 50, 0);
-		return ;
-	}
-	// check if user has authenticated.
-	if (!client->is_authenticated())
-	{
-		const std::string &errorMsg = "ERROR :You must authenticate with PASS first\r\n";
-		sendError(fd, errorMsg);
-		std::cout << "[WARN] Client " << fd << " tried to send NICK/USER before PASS\r\n";
-		return ;
-	}
-	// iterate to the second item in the list
-	// /NICK nickname
-	std::list<std::string>::const_iterator it = cmd_list.begin();
-	++it;
-	std::string newNick = *it;
-	// check if the nickname has a max of 9 characters
-
-	if (newNick.length() > 9)
-	{
-		sendError(fd, "ERROR :Nickname too long\r\n");
-		return;
-	}
-	std::string oldNick = client->getNick();
-	client->set_nick(newNick);
-	std::cout << "[NICK] " << oldNick << " changed to " << newNick << std::endl;
-	std::string nickReply = ":" + oldNick + "!" + client->getUserName() + "@" +
-	client->getHostName() + " NICK :" + newNick + CRLF;
-	sendReply(fd, nickReply);
-}
 
 void Server::sendWelcome(Client *client)
 {
@@ -316,313 +246,11 @@ void Server::sendWelcome(Client *client)
 	sendReply(fd, RPL_ENDOFMOTD(serverName, nick));
 }
 
-/**
- * Sets the user's nick, username and hostname
- */
-void Server::handleUser(int fd, std::list<std::string> cmd_list)
-{
-	Client *client = getClientByFd(fd);
-	if (!client)
-		return ;
-	if (cmd_list.size() != 5)
-	{
-		const std::string &errorMsg = "ERROR :USER command requires three argument\r\n";
-		sendError(fd, errorMsg);
-		return ;
-	}
-	// check if user has authenticated.
-	if (!client->is_authenticated())
-	{
-		const std::string &errorMsg = "ERROR :You must authenticate with PASS first\r\n";
-		sendError(fd, errorMsg);
-		std::cout << "[WARN] Client " << fd << " tried to send NICK/USER before PASS\n";
-		return ;
-	}
-	// iterate to the second item in the list
-	// /USER username something hostname
-	std::list<std::string>::const_iterator it = cmd_list.begin();
-	++it;
-	std::string username = *it;
-	//std::cout << second << std::endl;
-	client->set_username(username);
-	++it;++it;
-	std::string hostname = *it;
-	//std::cout << fourth << std::endl;
-
-	client->set_hostname(hostname);
-	std::cout << "[USER] : " << username << ", " << "[USER] : "<< hostname  << std::endl;
-	// check if all is set, return a message
-	if (!client->getNick().empty() && !client->getUserName().empty()
-		&& !client->getHostName().empty() && !client->is_registered())
-	{
-		client->register_client();
-		sendWelcome(client);
-	}
-}
-
-/**
- * Helps user join a channel. If channel is not created on server, this creates the new channel.
- * For now we create channels that does not require passwords
- */
-void Server::handleJoin(int fd, std::list<std::string> cmd_list)
-{
-	Client *client = getClientByFd(fd);
-	if (!client)
-		return ;
-
-	if (cmd_list.size() < 2 || cmd_list.size() > 3)
-	{
-		send(fd, "ERROR :JOIN command requires 1 or 2 arguments\r\n", 47, 0);
-		return ;
-	}
-	// extract the channel name
-	std::list<std::string>::const_iterator it = cmd_list.begin();
-	++it;
-	const std::string channelName = *it;
-	//check if there is a password provided by user
-	std::string password = "";
-	if (++it != cmd_list.end())
-		password = *it;
-	// iterate over _channels to search if the channel already exist
-	Channel *channel = getChannelByName(channelName);
-
-	// if the channel does not exit, create the channel
-	if (!channel)
-	{
-		// create a new channel with the name given
-		_channels.push_back(Channel(channelName, password));
-		// get a reference to the channel created
-		channel = &_channels.back();
-		// add the client as operator
-		channel->addOperator(client);
-		// server displays message to show new channel is created
-		std::cout << "[INFO] New channel " << channelName
-				  << " created by " << client->getNick() << "\r\n";
-		std::cout << channelName << " is a password protected channel" << "\r\n";
-		std::string clientList = channel->getClientList();
-		std::cout << "printing client list" << std::endl;
-		std::cout << clientList << std::endl;
-		// send all the message together to irssi
-		/**
-		 * Need to get the variables to avoid calling getters multiple times;
-		 */
-		sendReply(fd,
-		RPL_JOINMSG(client->getNick(), client->getUserName(), client->getHostName(), channelName) +
-		RPL_TOPIC(getName(), client->getNick(), channelName, channel->getTopic()) +
-		RPL_CREATIONTIME(getName(), client->getNick(), channelName, channel->getCreationTime()) +
-		RPL_NAMEREPLY(getName(), client->getNick(), channelName, clientList) +
-		RPL_ENDOFNAMES(getName(), client->getNick(), channelName));
-		// broadcast the join mesasge to all others except user
-		channel->broadcast(RPL_JOINMSG(client->getNick(), client->getUserName(), client->getHostName(), channelName), client);
-	}
-	else
-	{
-		// this channel already exist, add the client as member
-		// need to check if client is already a member
-		if (!channel->isMember(client))
-		{
-			// check if there is password and if the password provided is the same
-			if (channel->hasPassword() && channel->getPassword() != password)
-			{
-				std::cout << "Error: " << client->getNick() << " tried to join without a correct password" << std::endl;
-				sendError(fd, ERR_BADCHANNELKEY(getName(), client->getNick(), channelName));
-				return ;
-			}
-			channel->addMember(client);
-
-			std::cout << "[JOIN] " << client->getNick() << " joined " << channelName << "\n";
-			std::cout << "[USERS] " << channel->getClientList() << "\n";
-
-			std::string clientList = channel->getClientList();
-			std::cout << "printing client list" << std::endl;
-			std::cout << clientList << std::endl;
-			std::cout << "NAMEREPLY" << std::endl;
-			std::cout << RPL_NAMEREPLY(getName(), client->getNick(), channelName, clientList);
-			// send all the message together to irssi
-			sendReply(fd,
-				RPL_JOINMSG(client->getNick(), client->getUserName(), client->getHostName(), channelName) +
-				RPL_TOPIC(getName(), client->getNick(), channelName, channel->getTopic()) +
-				//RPL_CREATIONTIME(getName(), client->getNick(), channelName, channel->getCreationTime()) +
-				RPL_NAMEREPLY(getName(), client->getNick(), channelName, clientList) +
-				RPL_ENDOFNAMES(getName(), client->getNick(), channelName));
-			std::cout << RPL_JOINMSG(client->getNick(), client->getUserName(), client->getHostName(), channelName) << std::endl;
-			// broadcast the join mesasge to all others except user
-			channel->broadcast(RPL_JOINMSG(client->getNick(), client->getUserName(), client->getHostName(), channelName), client);
-		}
-	}
-}
-
-void	Server::handlePing(int fd, std::list<std::string> cmd_lst)
-{
-	if (cmd_lst.size() < 2)
-	{
-		sendError(fd, "ERROR :No ping argument\r\n");
-		return;
-	}
-	std::list<std::string>::iterator it = cmd_lst.begin();
-	++it;
-	std::string token = *it;
-	sendReply(fd, RPL_PONG(token));
-	std::cout << "[PING] Replied with: " << RPL_PONG(token);
-}
-
-
-
-void	Server::handlePart(int fd, std::list<std::string> cmd_list)
-{
-	Client *client = getClientByFd(fd);
-	if (!client)
-	{
-		std::cout << "[PART] No client found for fd: " << fd << std::endl;
-		return ;
-	}
-	// if there are only one param, throw error
-	if (cmd_list.size() == 1)
-	{
-		sendError(fd, ERR_NEEDMOREPARAMS(getName(), client->getNick(), cmd_list.front()));
-		std::cout << "[PART] Not enough parameters" << std::endl;
-		return ;
-	}
-	// check if channel exit
-	// extract the channel name
-	std::list<std::string>::const_iterator it = cmd_list.begin();
-	++it;
-	const std::string channelName = *it;
-	std::string reason = "";
-	if (cmd_list.size() > 2)
-	{
-		++it;
-		while (it != cmd_list.end())
-		{
-			if (!reason.empty())
-			{
-				reason += " ";
-			}
-			reason += *it;
-			++it;
-		}
-	}
-	std::cout << "Reason" << std::endl;
-	std::cout << reason << std::endl;
-	// iterate over _channels to search if the channel already exist
-	Channel *channel = getChannelByName(channelName);
-
-	if (!channel)
-	{
-		sendError(fd, ERR_NOSUCHCHANNEL(getName(), client->getNick(), channelName));
-		std::cout << "[PART] Channel " << channelName << " not found" << std::endl;
-		return;
-	}
-	if (!channel->isMember(client) && !channel->isOperator(client))
-	{
-		sendError(fd, ERR_NOTONCHANNEL(getName(), client->getNick(), channelName));
-		std::cout << "[PART] Client not in channel " << channelName << std::endl;
-		return;
-	}
-
-	// may need include the parting message
-
-	std::string partMsg = ":" + client->getPrefix() + " PART " + channelName;
-	if (!reason.empty())
-		partMsg += " " + reason;
-	else
-		partMsg + " :";
-	partMsg += "\r\n";
-
-	std::cout << "DEBUG PART MSG: [" << partMsg << "]" << std::endl;
-	sendReply(client->getFd(), partMsg);
-	channel->broadcast(partMsg, client);
-	// remove the member/operator from the channel;
-	channel->removeUser(client);
-
-	// after the user has been removed, broad the message to other users
-	// if there is no more users in the channel, remove it from the channel vector in the server
-	if (channel->getUsers().empty())
-	{
-		removeChannel(channel->getName());
-		sendError(fd, ERR_NOSUCHCHANNEL(getName(), client->getNick(), channelName));
-	}
-	else
-		sendError(fd, ERR_NOTONCHANNEL(getName(), client->getNick(), channelName));
-}
-
-void	Server::handlePrivmsg(int fd, std::list<std::string> cmd_list)
-{
-	Client *client = getClientByFd(fd);
-	if (!client)
-	{
-		std::cout << "[PRIVMSG] No client found for fd: " << fd << std::endl;
-		return ;
-	}
-	// if there are only one param, throw error
-	if (cmd_list.size() < 3)
-	{
-		sendError(fd, ERR_NEEDMOREPARAMS(getName(), client->getNick(), cmd_list.front()));
-		std::cout << "[PRIVMSG] Not enough parameters" << std::endl;
-		return ;
-	}
-	std::list<std::string>::iterator it = cmd_list.begin();
-	++it;
-	std::string target = *it++;
-	std::string message;
-	// add all the strings into message
-	while (it != cmd_list.end())
-	{
-		if (!message.empty())
-			message += " ";
-		message += *it++;
-	}
-	if (message[0] == ':')
-		message = message.substr(1);
-
-	if (target[0] == '#') {
-		std::cout << "[DEBUG] Message is for channel: " << target << std::endl;
-	} else {
-		std::cout << "[DEBUG] Message is for user: " << target << std::endl;
-	}
-
-	if (target[0] == '#')
-	{
-		Channel *channel = getChannelByName(target);
-		if (!channel)
-		{
-			sendError(fd, ERR_NOSUCHCHANNEL(getName(), client->getNick(), channel->getName()));
-			return ;
-		}
-		if (!channel->isMember(client))
-		{
-			sendError(fd, ERR_CANNOTSENDTOCHAN(getName(), client->getNick(), channel->getName()));
-			return ;
-		}
-		std::string out = ":" + client->getPrefix() + " PRIVMSG " + target + " :" + message + CRLF;
-		std::cout << out << std::endl;
-		channel->broadcast(out, client);
-	}
-	else
-	{
-		// find the target in the server _clients
-		Client* targetUser = getClientByNick(target);
-		if (!targetUser)
-		{
-			std::cerr << "[DEBUG] No such user: " << target << std::endl;
-			ERR_NOSUCHNICK(getName(), client->getNick(), target);
-			return ;
-		}
-		std::cout << "[DEBUG] Target nick: " << targetUser->getNick() << std::endl;
-		std::string out = ":" + client->getPrefix() + " PRIVMSG " + target + " :" + message + CRLF;
-		std::cout << out << std::endl;
-		// send the message to the target user
-		sendReply(targetUser->getFd(), out);
-	}
-}
-
-
-
 void	Server::execute_cmd(int fd, std::list<std::string> cmd_lst)
 {
 	Client* client = getClientByFd(fd);
 	const std::string &cmd = cmd_lst.front();
-	//const std::string &errorMsg = "ERROR :Unknown command\r\n";
+
 	if (cmd == "PASS")
 		// authenticate the user
 		handlePass(fd, cmd_lst);
